+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Filter internal adapters from channels tab and inline setup wizard
 ---
 
 ## [0.35.0] - 2026-04-09